backend/packages/cognitest-common/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cognitest-common"
version = "0.1.0"
description = "Common utilities for Cognitest-AI microservices"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pydantic-settings>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.0.1",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.28.0",
    "redis>=5.0.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.1.0",
    "google-generativeai>=0.5.0"
]
